fix: type assertion for config class
